Speed up packet reading and writing
This commit started by replacing some `.times` calls with `while` loops,
but then went a bit further by hard-coding the different int methods
instead of looping, doing ivar lookups, multiplying, etc.

Running the benchmark:

```
ruby --yjit bin/benchmark
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [x86_64-darwin23]
```

is now about 1.5x slower than trilogy instead of 1.8x slower.

+++ lib/nocturne/protocol.rb
@@ -8,8 +8,8 @@ class Nocturne
     end
 
     def self.read_error(packet)
-      packet.int
-      code = packet.int(2)
+      packet.int8
+      code = packet.int16
       packet.strn(1)
       packet.strn(5)
       message = packet.eof_str
